removed test from nodes file, created new test file for readonly
